+++ src/Cart/Contracts/Cart.php
@@ -24,50 +24,40 @@ interface Cart extends CheckoutSubject
      * Add an item to the cart (or adds the quantity if the product is already in the cart)
      *
      * @param Buyable $product Any Buyable object
-     * @param int $qty The quantity to add
+     * @param int|float $qty The quantity to add
      * @param array $params Additional parameters, eg. coupon code
      *
      * @return CartItem Returns the item object that has been created (or updated)
      */
-    public function addItem(Buyable $product, $qty = 1, $params = []): CartItem;
+    public function addItem(Buyable $product, int|float $qty = 1, array $params = []): CartItem;
 
     /**
      * Removes an item from the cart
-     *
-     * @param object|int    $item    Object: item or int = item id
      */
-    public function removeItem($item);
+    public function removeItem(CartItem $item): void;
 
     /**
      * Removes a product from the cart
-     *
-     * @param Buyable $product
      */
-    public function removeProduct(Buyable $product);
+    public function removeProduct(Buyable $product): void;
 
     /**
      * Clears the entire cart
      */
-    public function clear();
+    public function clear(): void;
 
     /**
      * Returns the number of items in the cart
-     *
-     * @return int
      */
-    public function itemCount();
+    public function itemCount(): int;
 
     /**
      * Returns the cart's associated user, or NULL
-     *
-     * @return Authenticatable|null
      */
-    public function getUser();
+    public function getUser(): ?Authenticatable;
 
     /**
-     * Set the user of the cart
-     *
-     * @param Authenticatable|integer $user User object or user id
+     * Set the user of the cart by passing a user object or user id
      */
-    public function setUser($user);
+    public function setUser(Authenticatable|int|string|null $user): void;
 }