+++ src/Cart/Contracts/CartManager.php
@@ -18,58 +18,48 @@ interface CartManager extends Cart
 {
     /**
      * Returns whether a cart exists for this user or session
-     *
-     * @return bool
      */
-    public function exists();
+    public function exists(): bool;
 
     /**
      * Returns true if neither the session nor the user has a cart
-     *
-     * @return bool
      */
-    public function doesNotExist();
+    public function doesNotExist(): bool;
 
     /**
      * Returns the encapsulated ActiveRecord model or null if not exists
-     *
-     * @return Cart|null
      */
-    public function model();
+    public function model(): ?Cart;
 
     /**
      * Returns true if the cart is empty (doesn't contain items)
-     *
-     * @return bool
      */
-    public function isEmpty();
+    public function isEmpty(): bool;
 
     /**
      * Returns true if the cart is not empty (contains items)
-     *
-     * @return bool
      */
-    public function isNotEmpty();
+    public function isNotEmpty(): bool;
 
     /**
      * Completely destroys the cart: removes all related models (cart, item, etc) from the DB
      */
-    public function destroy();
+    public function destroy(): void;
 
     /**
      * Creates a new cart
      *
      * @param bool $forceCreateIfExists Creates a new cart even if there's an existing one
      */
-    public function create($forceCreateIfExists = false);
+    public function create(bool $forceCreateIfExists = false): void;
 
     /**
      * Dissociates a user from a cart
      */
-    public function removeUser();
+    public function removeUser(): void;
 
     /**
      * Forget the cart for the session, but keep it
      */
-    public function forget();
+    public function forget(): void;
 }